Added functionality to insert Product into the database from AddProductFragment.

+++ app/src/main/java/com/github/albertobf/grocerycompanion/model/Product.kt
@@ -13,7 +13,7 @@ data class Product(
     @PrimaryKey(autoGenerate = true)
     var id: Long = 0L,
     val name: String,
-    val size: Float,
+    val size: Int,
     @Embedded(prefix = "sizetype_")
     val sizeType: SizeType
 )
